+++ src/main/java/lotto/lotto/LottoMachine.java
@@ -1,5 +1,7 @@
 package lotto.lotto;
 
+import lotto.lotto.lottonumber.LottoNumbers;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
