+++ src/test/java/lotto/LottoMachineTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.lotto.LottoMachine;
-import lotto.lotto.LottoNumbers;
+import lotto.lotto.lottonumber.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
