+++ src/test/java/lotto/LottoNumberTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.lotto.LottoNumber;
+import lotto.lotto.lottonumber.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
