+++ src/test/java/lotto/LottoNumbersTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.lotto.LottoNumbers;
+import lotto.lotto.lottonumber.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
