+++ src/test/java/lotto/LottosTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.lotto.Lotto;
-import lotto.lotto.LottoNumbers;
+import lotto.lotto.lottonumber.LottoNumbers;
 import lotto.lotto.Lottos;
 import lotto.result.LottoResults;
 import org.junit.jupiter.api.DisplayName;
